+++ src/accordion/__tests__/AccordionState.test.ts
@@ -14,31 +14,10 @@ function render({
     .result;
 }
 
-test("Accordion: initial state", () => {
-  const result = render();
+describe("useAccordionState", () => {
+  test("initial state", () => {
+    const result = render();
 
-  expect(result.current).toMatchInlineSnapshot(`
-    Object {
-      "allowMultiple": false,
-      "allowToggle": false,
-      "baseId": "base",
-      "currentId": undefined,
-      "groups": Array [],
-      "items": Array [],
-      "loop": false,
-      "manual": true,
-      "orientation": "vertical",
-      "panels": Array [],
-      "rtl": false,
-      "selectedId": null,
-      "unstable_angular": false,
-      "unstable_hasActiveWidget": false,
-      "unstable_idCountRef": Object {
-        "current": 0,
-      },
-      "unstable_moves": 0,
-      "unstable_virtual": false,
-      "wrap": false,
-    }
-  `);
+    expect(result.current).toMatchSnapshot();
+  });
 });