+++ src/toast/__tests__/Toast.test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { axe, render, click, act } from "reakit-test-utils";
+import { axe, render, click, act, screen } from "reakit-test-utils";
 
 import { useToast, ToastProvider } from "..";
 
@@ -29,21 +29,21 @@ const Demo = () => {
           });
         }}
       >
-        Error
+        Error button
       </button>
       <button
         onClick={() => {
           showToast({ type: "success", content: "This is success" });
         }}
       >
-        Success
+        Success button
       </button>
       <button
         onClick={() => {
           showToast({ type: "warning", content: "This is warning" });
         }}
       >
-        Warning
+        Warning button
       </button>
       <button
         onClick={() => {
@@ -56,7 +56,7 @@ const Demo = () => {
           });
         }}
       >
-        Custom
+        Custom button
       </button>
     </div>
   );
@@ -70,7 +70,7 @@ const ToastComp: React.FC = () => {
         error: ({ hideToast, content, id }) => {
           return (
             <div
-              data-testid="error"
+              data-testid="error-toast"
               className="toast"
               style={{ backgroundColor: "#f02c2d" }}
             >
@@ -111,16 +111,16 @@ describe("Toast", () => {
         <div>
           <div>
             <button>
-              Error
+              Error button
             </button>
             <button>
-              Success
+              Success button
             </button>
             <button>
-              Warning
+              Warning button
             </button>
             <button>
-              Custom
+              Custom button
             </button>
           </div>
         </div>
@@ -129,39 +129,42 @@ describe("Toast", () => {
   });
 
   it("toast should popup to the screen after click", () => {
-    const { getByText: text, getByTestId: id } = render(<ToastComp />);
+    render(<ToastComp />);
 
-    expect(text("Error")).toBeInTheDocument();
+    const errorBtn = screen.getByText("Error button");
+
+    expect(errorBtn).toBeInTheDocument();
 
     act(() => {
-      click(text("Error"));
+      click(errorBtn);
     });
 
-    expect(id("error")).toHaveTextContent("This is error");
+    expect(screen.getByTestId("error-toast")).toHaveTextContent(
+      "This is error",
+    );
   });
 
   it("should be hideToastd after clicking close button", () => {
-    const {
-      getByText: text,
-      getByTestId: getId,
-      queryByTestId: queryId,
-    } = render(<ToastComp />);
+    render(<ToastComp />);
+    const errorBtn = screen.getByText("Error button");
 
-    expect(text("Error")).toBeInTheDocument();
+    expect(errorBtn).toBeInTheDocument();
 
     // add first
     act(() => {
-      click(text("Error"));
+      click(errorBtn);
     });
 
-    expect(getId("error")).toHaveTextContent("This is error");
+    expect(screen.getByTestId("error-toast")).toHaveTextContent(
+      "This is error",
+    );
 
     // let hideToast now
     act(() => {
-      click(getId("error-close"));
+      click(screen.getByTestId("error-close"));
     });
 
-    expect(queryId("error")).not.toBeInTheDocument();
+    expect(screen.queryByTestId("error-toast")).not.toBeInTheDocument();
   });
 
   test("Toast renders with no a11y violations", async () => {