+++ src/utils/test-utils.ts
@@ -1,21 +1,13 @@
-import { Matcher } from "@testing-library/react";
-
-export const isEndSelection = (getByLabelText: any, label: Matcher) => {
-  expect(getByLabelText(label).parentElement).toHaveAttribute(
-    "data-is-selection-end",
-  );
+export const isEndSelection = (element: HTMLElement) => {
+  expect(element.parentElement).toHaveAttribute("data-is-selection-end");
 };
 
-export const isStartSelection = (getByLabelText: any, label: Matcher) => {
-  expect(getByLabelText(label).parentElement).toHaveAttribute(
-    "data-is-selection-start",
-  );
+export const isStartSelection = (element: HTMLElement) => {
+  expect(element.parentElement).toHaveAttribute("data-is-selection-start");
 };
 
-export const isInSelectionRange = (getByLabelText: any, label: Matcher) => {
-  expect(getByLabelText(label).parentElement).toHaveAttribute(
-    "data-is-range-selection",
-  );
+export const isInSelectionRange = (element: HTMLElement) => {
+  expect(element.parentElement).toHaveAttribute("data-is-range-selection");
 };
 
 export const repeat = (cb: Function, times: number) => {